fix route listing project and fix responsive landx-cuan
