+++ src/components/project-detail/detailProjects.scss
